+++ R/zzz.R
@@ -9,6 +9,27 @@
   s3_register("generics::required_pkgs", "cluster_fit")
   s3_register("generics::required_pkgs", "cluster_spec")
 
+  s3_register("butcher::axe_call", "cluster_fit")
+  s3_register("butcher::axe_ctrl", "cluster_fit")
+  s3_register("butcher::axe_data", "cluster_fit")
+  s3_register("butcher::axe_env", "cluster_fit")
+  s3_register("butcher::axe_fitted", "cluster_fit")
+
+  s3_register("butcher::axe_data", "kmeans")
+  s3_register("butcher::axe_fitted", "kmeans")
+
+  s3_register("butcher::axe_data", "KMeansCluster")
+  s3_register("butcher::axe_fitted", "KMeansCluster")
+
+  s3_register("butcher::axe_data", "kproto")
+  s3_register("butcher::axe_fitted", "kproto")
+
+  s3_register("butcher::axe_data", "kmodes")
+  s3_register("butcher::axe_fitted", "kmodes")
+
+  s3_register("butcher::axe_data", "Mclust")
+  s3_register("butcher::axe_fitted", "Mclust")
+
   ns <- rlang::ns_env("tidyclust")
   makeActiveBinding(
     "tidyclust_color",